+++ src/utils/UniqueArray.sol
@@ -23,7 +23,7 @@ function _requireNoDuplicates(address[] calldata arr) pure {
     }
 }
 
-function _requireNoNegatives(int88[] memory vals) pure {
+function _requireNoNegatives(int256[] memory vals) pure {
     uint256 arrLength = vals.length;
 
     for (uint i; i < arrLength; i++) {
